1.3.0 UPDATE

+++ app/plugins/resend.js
@@ -1,19 +1,4 @@
-const text = {
-    "unsendNoBody": "• %1 has unsend a message",
-    "unsendNoBodyButAttach": "• %1 has unsend %2 attachments",
-    "unsendBodyAndAttach": "• %1 has unsend a message:\n%2\n• With %3 attachments:",
-    "unsendBody": "• %1 has unsend a message:\n%2"
-}
-
-const getText = (key, ...args) => {
-    if (text[key]) {
-        return text[key].replace(/%(\d+)/g, (match, p1) => args[p1 - 1]);
-    }
-    return key;
-}
-
-import { writeFileSync, createReadStream, unlinkSync } from 'fs';
-import get from 'axios';
+import { writeFileSync, createReadStream, unlinkSync } from "fs";
 
 const exts = {
     "photo": ".jpg",
@@ -21,9 +6,9 @@ const exts = {
     "audio": ".mp3",
     "animated_image": ".gif",
     "share": ".jpg",
-    "file": ''
+    "file": ""
 }
-const cachePath = process.cwd() + '/plugins/cache/';
+const cachePath = process.cwd() + "/plugins/cache/";
 
 export default async function ({ api, event, db, controllers }) {
     const { Threads, Users } = controllers;
@@ -35,26 +20,34 @@ export default async function ({ api, event, db, controllers }) {
     if (threadData.resend == false) return;
 
     if (!client.data.message.some(item => item.msgID == messageID)) return;
-    var getMsg = client.data.message.find(item => item.msgID == messageID);
+    let getMsg = client.data.message.find(item => item.msgID == messageID);
     let name = await Users.getName(senderID);
-    var messageObject = {},
-        msg = '';
+    let messageObject = {},
+        msg = "";
 
     if (!getMsg.body && getMsg.attachment.length == 0) {
-        return api.sendMessage(getText('unsendNoBody', name), threadID, messageID);
+        return api.sendMessage(getLang("plugins.resend.noMessage", { author: name }), threadID);
     }
     if (getMsg.attachment.length > 0) {
-        if (getMsg.body) msg = getText('unsendBodyAndAttach', name, getMsg.body, getMsg.attachment.length);
-        else msg = getText('unsendNoBodyButAttach', name, getMsg.attachment.length);
-        var attachment = [];
-        var filePath = [];
+        if (getMsg.body) msg = getLang("plugins.resend.unsendMessageAndAttach", {
+            author: name,
+            message: getMsg.body,
+            attachmentsCount: getMsg.attachment.length
+        })
+        else msg = getLang("plugins.resend.unsendNoMessageButAttach", {
+            author: name,
+            attachmentsCount: getMsg.attachment.length
+        })
+        let attachment = [];
+        let filePath = [];
         for (let i = 0; i < getMsg.attachment.length; i++) {
             try {
                 const { fileName } = getMsg.attachment[i] || `${Date.now()}_${senderID}_${i}`;
-                const ext = exts[getMsg.attachment[i].type] || '';
+                const ext = exts[getMsg.attachment[i].type] || "";
+                if (fileName + ext == ".gitkeep" || fileName + ext == 'README.txt') continue;
                 const savePath = `${cachePath}${fileName + ext}`;
-                const mediaBuffer = (await get(getMsg.attachment[i].url, { responseType: 'arraybuffer' })).data;
-                writeFileSync(savePath, Buffer.from(mediaBuffer, 'utf-8'));
+                const mediaBuffer = (await get(getMsg.attachment[i].url, { responseType: "arraybuffer" })).data;
+                writeFileSync(savePath, Buffer.from(mediaBuffer, "utf-8"));
                 attachment.push(createReadStream(savePath));
                 filePath.push(savePath);
             } catch (err) {
@@ -66,7 +59,10 @@ export default async function ({ api, event, db, controllers }) {
             attachment: attachment
         }
     } else if (getMsg.body) {
-        messageObject.body = getText('unsendBody', name, getMsg.body);
+        messageObject.body = getLang("plugins.resend.unsendMessage", {
+            author: name,
+            message: getMsg.body
+        });
     }
     if (Object.keys(messageObject).length > 0) {
         return api.sendMessage(messageObject, threadID, () => {
